fix: test should be able to run on the pipeline
