+++ .gitignore
@@ -20,5 +20,4 @@ Cargo.lock
 
 .idea
 
-/data/large
 /data/faulty
